backend/requirements.txt
fastapi
uvicorn
snowflake-connector-python
pydantic
pydantic-settings
google-generativeai
python-multipart
aiofiles
python-dotenv
cryptography
tenacity
geopy
